requirements.txt
pillow
datasets
tqdm
pandas
matplotlib
tensorflow
scipy
ultralytics
streamlit
streamlit-webrtc 
opencv-python